Well formatted chat list. Still we can't respond to chats yet
There is a field to respond to chats.
We should have a loggedInUser to track who is logged in.
Delete useless code in Renderer
We should only refresh if the new incoming chat list is different from what we have

TODOfu - Reply chats, Load chat window when someone is clicked on the onlinelist form